+++ CHANGELOG.md
@@ -28,6 +28,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Disable operator metrics by default. This removes a dependency on an external container.
 - Dependency on cert-manager for initial deployment.
 
+### Fixed
+
+- A crash caused by insufficient permissions on the LINSTOR Controller.
+
 ## [v2.1.1] - 2023-05-24
 
 ### Added